[feature #16797166]Move constant file to utils folder
